requirements.txt
openai
pyyaml
pymysql
fdb
fastapi
uvicorn
jinja2